+++ src/c++/lib/alignment/SingleRefAlignerShared.hh
@@ -62,7 +62,13 @@ struct SingleRefAlignerBase : public AlignerBase<ScoreType>
 {
     SingleRefAlignerBase(
         const AlignmentScores<ScoreType>& scores) :
-        AlignerBase<ScoreType>(scores)
+        AlignerBase<ScoreType>(scores,scores)
+    {}
+
+    SingleRefAlignerBase(
+            const AlignmentScores<ScoreType>& scores,
+            const AlignmentScores<ScoreType>& contigFilterScores) :
+            AlignerBase<ScoreType>(scores,contigFilterScores)
     {}
 
 protected: